maintain: rename package `ui` to `view`, `logic` to `domain`
