+++ src/main/java/race/domain/Car.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 import java.util.Objects;
