+++ src/main/java/race/domain/LapScores.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 import java.util.List;
 import java.util.Objects;
