+++ src/main/java/race/domain/RandomEngine.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 import java.util.Random;
