+++ src/main/java/race/domain/ResultEmitter.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 public interface ResultEmitter {
     void emitScores(LapScores lapScores);
