+++ src/test/java/race/domain/CarTest.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
