+++ src/test/java/race/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package race.logic;
+package race.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
